Update FreeRTOS example
