--- examples/freertos/software/FreeRTOS.cmake
@@ -1,30 +1,0 @@
-# -----------------------------------------------------------------------------
-# Copyright (c) 2020-2024 RISC-V Steel contributors
-#
-# This work is licensed under the MIT License, see LICENSE file for details.
-# SPDX-License-Identifier: MIT
-# -----------------------------------------------------------------------------
-
-include (FetchContent)
-
-FetchContent_Declare( freertos_kernel
-  GIT_REPOSITORY https://github.com/FreeRTOS/FreeRTOS-Kernel.git
-  GIT_TAG        V11.0.0
-)
-
-add_library(freertos_config INTERFACE)
-
-target_include_directories(freertos_config SYSTEM
-  INTERFACE
-    include
-)
-
-target_compile_definitions(freertos_config
-  INTERFACE
-    projCOVERAGE_TEST=0
-)
-
-set(FREERTOS_PORT "GCC_RISC_V_GENERIC" CACHE STRING "" FORCE)
-set(FREERTOS_RISCV_EXTENSION RISCV_no_extensions CACHE STRING "")
-
-FetchContent_MakeAvailable(freertos_kernel)
